Various updates

+++ src/usr/local/kobocloud/udev_mount.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 #load config
-. `dirname $0`/config.sh
+. $(dirname $0)/config.sh
 
 #create work dirs
 [ ! -e "$Logs" ] && mkdir -p "$Logs" >/dev/null 2>&1
@@ -8,9 +8,13 @@
 [ ! -e "$SD" ] && mkdir -p "$SD" >/dev/null 2>&1
 
 if [ ! -e $UserConfig ]; then
-  echo "# Add your URLs to this file" > $UserConfig
-  echo "# Remove the # from the following line to uninstall KoboCloud" >> $UserConfig
-  echo "#UNINSTALL" >> $UserConfig
+  if [ -e $ConfigFile ]; then
+    cp $ConfigFile $UserConfig
+  else
+    echo "# Add your URLs to this file" > $UserConfig
+    echo "# Remove the # from the following line to uninstall KoboCloud" >> $UserConfig
+    echo "#UNINSTALL" >> $UserConfig
+  fi
 fi
 
 #bind mount to subfolder of SD card on reboot
